+++ lib/glimmer/libui/control_proxy/column/background_color_column_proxy.rb
@@ -31,6 +31,10 @@ module Glimmer
         # Follows the Proxy Design Pattern
         class BackgroundColorColumnProxy < ControlProxy
           include Column
+          
+          def name
+            'Background Color'
+          end
         end
       end
     end